Chinese (Simplified) translations
I hope you could take some time and check this to make sure nothing is wrong with and part of original coding. Also, is it possible to change place of the Minecraft version in the text "Read to play Minecraft X.XX"? I can't find a perfect suitable text for it.
Also, what should I do with the account registered in the account created in SKLauncher website? Can I login with that?