Create sklmessages_bn_IN.properties
Though I haven't found any code for Bengali (INDIA) language in the website for language codes referred in "Translating SKlauncher"  guide, but then also with the help of the Hindi (INDIA) translation, I made this for the 'only Bengali' users. This is my contribution... I hope if it helps. 
Thank you...
RupamSaha789#Progamrz50op